Added support for direct connection to the libvirt daemon on compute nodes in OpenStack clouds with KVM-based virtualization
Fixed a bug that prevented VMs from being properly removed in case of VApp
deployment failure in OpenStack.

Added a new method to check if a VM completed its boot process: "run_command".
With this method, CB will run a command and consider that the VM is booted in
case the exit status is 0.