Several bugfixes and improvements
1) The checking for passwordless ssh access to the JUMP_HOST (when configured) happens only
after the connection to the cloud (and creation of the appropriate SSH key pair there).
This manner a user can right after the cloud attachment failure, just create a new VM to be
the JUMP_HOST, passing the appropriate public ssh key (cbtool_rsa) by default.

2) Now, by default, logs go to /var/log/cloudbench. Each one of the multiple log files are
prefixed by the USERNAME that launched cloudbench (e.g., cbuser_operations.log), a measure
to allow multiple user to run CB on the same host simultaneously.

3) When an non-existing cloud model is indicated during a cldattach (e.g., "cldattach xyz
TESTCLOUD"), provide an appropriate error message.

4) The default cloud configuration file (i.e., USERNAME_cloud_definitions.txt) can now
reside within a subdirectory inside the configs directory.

5) During the post attachment operations on the VM, a new separate step, before the copy
of the code is rsynced and the post-boot script is run remotely, is executed. In this step
CB just checks to make sure that the ssh is really inject (or already inside) the VM by
remotely running "/bin/true"

6) Fixes an error during the CB startup where the private rsyslog configuration file was
deleted, but the rsyslogd process is still running.

7) The configuration parser now supports nested INCLUDE statements