Multiple changes:
1) Attributes in configuration files now support multi-line values by using the "+=" sign.
For instance an excerpt like the below is now perfectly legal:

...
MY_MULTILINE_ATTRIBUTE = first line
MY_MULTILINE_ATTRIBUTE += second line
MY_MULTILINE_ATTRIBUTE += one more

2) In an effort to make CB more "self-documenting", and the workloads and cloud adpaters
more self-contained, their specific configuration files are being moved. For instance,
gone are the *_application_instances.txt files in configs/templates/ dir. Now, each
Virtual Application type in scripts/<vapp type>/ has a virtual_application.txt file that
contains all the configuration for that particular VApp. The configurations now include a
brief text describing the Virtual Application, shown through the "typeshow <vapptype>"
command.

3) Record the time zone/time difference between VMs and the CB Orchestrator Node (ON).
This information is highly relevant for at least two situations: a) CB ON and VMs just
happen to be actually located in different geographical regions or b) the timezone on the
VMs is not properly set. In CB, the python call (time.time()) is used to create timestamps.
This call is supposedly *defined* to be 1970, 1, 1 in UTC, but in at least one scenario
it could be experimentally verified to be using the localtime. In order to provide the
experimenter with the correct information to allow him to fix any potential errors (due to
timezone shifting) during the production of plots, now the "utc_offset_delta" is provided.
It is provided both as an attribute on the Object Store and as a column in the .csv files
extracted from the metric store.

4) In cb_common.sh, Ganglia (gmond) start/stop operations are split in their own separate
functions, making it easier to restart gmond during the Virtual Application's lifetime.