Changes:
1) (For OpenStack clouds) During initial cloud attachment, if the value of the VM_DEFAULTS
attribute JUMP_HOST is "$True", CB will create a VM named "cb_jumphost", attach
a floating IP to it, and then automatically use it as the jumphost. If, instead of
"$True", the value of the JUMP_HOST attribute is an already existing IP address, then
the creation of the "jumphost" VM will be bypassed.

2) Added proper description (plus additional code cleanup) for Virtual Workloads
cassandra_ycsb, coremark, ddgen, filebench, hadoop and nullworkload.

3) Introduced a new regression test targetting "real" Virtual Application instances
on a "real" (OpenStack) cloud. Up until this point, all regression tests only covered
a simulated cloud, and thus no actual testing of the Virtual Application load generating
code was performed. With this commit, all Virtual Application instances (starting
by the ones described in #2) will be tested on a real OpenStack cloud before commit.