Improvements and bugfixes
1) The cbssh.sh utility, found on cbtool/util/, can be used to obtain an ssh prompt on
a VM directly. For instance "~/cbtool/util/cbssh.sh vm_3" will automatically discover the
VM's IP address, login, and (if needed) the name of the Jumphost used to access the VM,
and issue the command accordingly.

2) Improvements on the "typeshow" command. Now the output of this command is divided in
four sections: brief description, general mandatory attributes, AI-specific mandatory
attributes, AI-specific optional attributes.

3) In an attempt to solve an ongoing problem with OpenStack clouds under stress (i.e.,
with hundreds or thousands of VM being deployed), CB will be more tolerant with errors
during "get" or "list" API calls. What seems to be happening is, during a VM deployment,
when CB is polling the Cloud in order to get the instance state, sometimes the "get" (or
"list" API calls return an error). Most of these errors seem to be transient, and CB will
now use a per-VM "get"/"list" error counter (default value is 10) before deciding to
consider a VM failed, and starting a rollback of the Virtual-Application deployment.

4) Several improvements for Cassandra YCSB scripts. First of all, the Virtual Application
deployment is considered complete only after *all* nodes are joined to the cluster (the
utility "nodetool" is employed to this end). Second, we now have a truly idempotent
"restart node" and "restart seed" scripts. At the end of load id execution, these scripts
are re-run, and seeds and nodes are restarted accordingly. The idea here is to prevent
problems stemming from nodes or seeds suddenly dying in the middle of an experiment.