Multiple changes
1 -  Bugfixes for installer in Ubuntu 14.04, and for pymongo 3.0

2 - A "file store" (rsyncd server) is automatically started whenever CB is executed.
This allows VMs to pull/push files from the Orchestrator. New Global Object [FILESTORE].

3 - If attribute "START_SERVER" is $True in the new global [VPN]. An OpenVPN server is
automatically configured and started. Default value for START_SERVER is $False.

4 - Ability to select between neutron and nova python clients when dealing with network
discovery (and network id determination) in OpenStack Clouds.

5 - Tested the creation of multi-home VMs in OpenStack. From now, by default, the jumpbox
will always be connected to all networks in case of OpenStack.

6 - Support for cloud-init. If the attribute "USERDATA" in a VM (or in the Global Object
[VM_DEFAULTS] is set to $True, then a script that will attempt to pull a copy of the code
from the Orchestrator during but will be sent to the VM. This script also publishes a
message at the end of the execution. This mechanism is also use to make VMs accessible
through CB's new VPN server (with the appropriate parameters set). A value of $False (which
is actually the default) for this USERDATA attribute will disable the use of cloud-init
altogether, while any other value will be interpreted as a file name from whose contents
will be passed to the VM through cloud-init.

7 - The function report_app_metrics in shell/common/cb_common.py will now automatically
report accumulated values for errors (application or sla runtime/provision) and averages
for all other metrics. For instance a metric like app_errors will automatically generate
also an “acc” attribute, while a metric like app_bandwidth will automatically generate also
“avg”, “max” and “min” attributes.

8 - The command "monextract" can now be called from an API. It will dump the csv files
on the Orchestrator node, and return the rsync-accessible URL where the files can be
downloaded from.

9 - The "monextract all" can now receive experiment ID as one of the parameters.

10 - If the attribute "EXPID_CHANGE_RESTART" is set to $True in the Global Object [LOGSTORE],
it will cause the flushing of the Log Store whenever the expid is changed.

11 - Multiple Options for an Orchestrator outside the cloud
      Option 1: Flat or Provider Network
               [VM_DEFAULTS]
               PROV_NETNAME=<flat network name>
               RUN_NETNAME=<flat network name>

      Option 2: Floating IPs
               [VM_DEFAULTS]
               USE_FLOATING_IP = $True

      Option 3: Jump Box
                [VM_DEFAULTS]
                USE_JUMPHOST = $True
                Note: This will auto set CREATE_JUMPHOST = $True

      Option 4: VPN
                 [VM_DEFAULTS]
                 USE_VPN_IP = $True

                 Note: This will auto set START_SERVER = $True in the Global Object [VPN]