Big update, with lots of features and optimizations/fixes.
Features:

- First implementation of multi-tenant support. Still very clumsy, it assumes that tenants
were already created.

- The OpenStack cloud adapter now performs a full service discovery on the cloud, producing
a per-host list of (Nova, Cinder and Neutron) services. This will be instrumental for the
appropriate behavior of the Fault Injection Request Submitter (FIRS).

- A new metric, app_quiescent_time is now reported by default for
every Virtual Application. This metric expresses the time between consecutive executions
of the load generating process, and thus can be used to estimate the actual VApp's duty
cycle.

- Added two new workloads: nuttcp and xping (for now, it simply uses "ping").

- For iperf, netperf, nuttcp and xping, a new VApp-specific modifier parameter, EXTERNAL_TARGET
is introduced. With this parameter, it is now possible to execute "north-south" network
benchmarks. If EXTERNAL_TARGET is anything other than the default "none", then emitter/client
will direct the flow to this target, instead of its counterpart VM inside the VApp.

- Expanded the set of application-specific options for netperf and iperf.

- New API/CLI operation: monpurge, which takes an experiment id and purges all data
referring to that experiment from the Metric Store.

- Updated fault scenarios to include neutron (instead of quantum :-) services.

Optimizations/Fixes:

- Fixes for multi-networking deployments.

- When switching experiment ids, make sure that all Hosts are re-registered on the Metric
Store with the new experiment id.

- Yet another attempt to solve problems with OpenStack and mass-deletion of VMs.

- Added two new API calls/CLI commands, "appload" and "appunload". These options
are just shortcuts to "cldalter" commands that disable the execution of the load manager.

- Fixed a problem with fault injection scenarios.

- Updated iperf UDP in order to properly capture Jitter and Loss. In case of multiple
flows, the Jitter is averaged over all flows.

- Updates and fixes for automated installer on RHEL 7.0.

- The API expid now returns also a list of experiments ids found on the metric store.

- Added utility script to import all VM images into a (OpenStack) cloud.

- Added the (OpenStack) specific parameter "use_cinderclient", to allow CB to establish
communication with Clouds without a functional Cinder service.