Improvements and bugfixes 1 - One more attempt to make VPN support more robust and cloud agnostic. A new parameter, VPN_ONLY is introduced, as part of the Global Object [VM_DEFAULTS]. Togethere with the parameter USE_VPN_IP and the (mandatory in case of VPN) USERDATA, the following behaviors can be induced:
USERDATA = $True
Will cause the generation of cloud-init script that simply pulls (rsync, etc.) a copy of
the code from the orchestrator and also publishes a message indicating the end of the boot.

USE_VPN_IP = $True
The cloud-init script will FIRST execute the OpenVPN client to connect to the VPN server,
and then execute the same steps as a). However all VM initiated traffic to the Stores (in
 the Orchestrator) is done through "normal", "outside of the cloud" IP addresses.

ONLY_VPN = $True
Same as b), but now all VM initiated traffic to the Stores (in the Orchestrator) is through
 OpenVPN-assigned IP addresses.

2 - OpenVPN-assigned VM IP address is now collected on the shared functions for cloud adapters.

3 - Implemented support for YCSB with both Cassandra 2.2 (using cqlsh), while keeping the
support for Cassandra 1 - 2.1.X (using cassandra-cli).

4 - Bugfixes for the API (ganglia Host OS usage metrics was not being properly displayed).

5 - In an attempt to deal with freezes during the execution of generic application
instance scripts, the command "sudo ntpd -qg" is now wrapped around a time-bound
python executor.