One-line VPN bugfix.
A previous commit uses the OpenVPN management port to "discover" when the local VPN IP address
changes (very useful as your laptop opens and closes).

Sometimes multiple IP changes are reflected in the log file, so we need to search the log
backwards to find the most recent IP to use instead of the oldest one.