Make barriers work
This commit makes barriers functionally correct: Until now, each AI participating in the barrier very often hit a race condition where they might subscribe to the pub/sub channel too late -- the publish message would be sent before
one or more had a chance to subscribe to the channel. This is solved by forcing all AIs to lock access to the counter in python and to further require them to all both decrement *and* increment the counter. Decrementing the counter
back to zero indicates to the "last" AI that it's safe to finally publish the "go" message, ensuring that all participating AIs have actually subscribed before the message goes out.

There is an open problem, however: The counter itself: If you destroy a bunch of AIs in the middle of a barrier, the counter gets left hanging at a non-zero value. The next time you fire off a bunch of AIs, the counter doesn't get reset to zero.

In order for this process to be repeatable, we would need a new command: Something along the lines of "barrier_reset" or something like that (name is not important) which sets the counter to zero.

I haven't implemented that command. If someone actually decides to use barriers regularly (or perhaps myself), I'll come back and implement it later.