POST-FREEZE: Libcloud Super Class!
This is a major commit for libcloud-based adapters: This creates a super class
around libvirt so that new libvirt-based clouds can be added with the absolute
minimalest of changes.

We've introduced a base function named "pre_vmcreate" which allows special cases,
like VCloud to tweak things if necessary.

Obviously, I don't have access to VCloud, so I need someone else to test it for me.

This should be a huge step forward for other clouds and get the ball rolling much
faster.