Important changes to improve CB's usability and "debuggability" for first time users.
1) First of all, we will guide the new user to the process of connecting to a cloud without
any of the required pre-configured images.
The first message that will appear on a during the initial connection is the following:

 ATTENTION! Given that none of the pre-configured images listed on CBTOOL's configuration file were detected on this cloud, we need to start by testing the ability to create instances using unconfigured images. Start by executing "vmattach check:<IMAGEID OF YOUR CHOICE>" on the CLI

At this point, the user can try to attach a single instance to this cloud, skipping any
attempts to actually establish the network connectivity between Orchestrator and instance.

e.g. "vmattach check:1373563"

If the instance creation is successful, the following message will appear:

 The ability to create instances using unconfigured images was successfully verified! We may now proceed to check the ability to connect (via ssh) to instances by executing "vmattach check:<IMAGEID OF YOUR CHOICE>:<USERNAME FOR LOGIN>" on the CLI.

Then the user can proceed to test the instance regarding: a) network reachability
(from the Orchestrator), and also SSH pubkey injection.

e.g. "vmattach check:1373563:ubuntu"

If this step is succesful, the following message will be displayed:

 The ability to create (and connect via ssh to) instances using unconfigured images was successfully verified. At this point, use this running instance as a base, and go straight to STEP 1 on https://github.com/ibmcb/cbtool/wiki/HOWTO:-Preparing-a-VM-to-be-used-with-CBTOOL-on-a-real-cloud

This behavior was implemented for the following clouds: SoftLayer, Amazon EC2, OpenStack,
Docker, and Google Compute Engine.
To do: VMWare, Digital Ocean, "pure" Libvirt (PLM, and FTC), SmartCloudProvider, CloudStack

2) For Amazon EC2 and SoftLayer clouds, allow an image name, instead of an image ID, to be
specified on "role" specification (the name gets automatically translated - internally -
into an ID).

3) Whenever the CLI/API vmattach is used, attempt to check, from the instance, if the
Object, Metric and Log store are reachable (on the Orchestrator).

4) Bugfixes for the SoftLayer Cloud adapter.

5) Allow JAVA_HOME and HADOOP_HOME to be set as Application Instance parameters for YCSB
(Cassandra, Redis and MongoDB) and Hadoop (Giraph too). In case the specified path is not
 present, a reasonable attempt will be made to locate it on the instance's disk.