Additional (small) changes and bugfixes on top the "walktrhough mode" PR.
1) Make sure that, for Kubernetes, VMs can be mapped to (single-container) Pods, ReplicaSets
and Deployments (this behavior is controlled by the parameter "ABSTRACTION" under [VM_DEFAULTS : KUB_CLOUDCONFIG]

2) Add a generic support for detailed timing breakdown annotation for events. This was implemented
for OpenStack, Docker (PDM) and Kubernetes adapters, resulting in a extra set of measurements
detailing the time spent in operations such as authenticating with the Cloud, searching for images, etc.
properly exported on a CSV file.

3) Several small bugfixes (e.g., API call/CLI command "waituntil")