Ability to automatically build qcow2 (kvm/qemu) images
This commit brings the ability to automatically create kvm-qemu images
(both Orchestrator and Workloads). It relies heavily on a recent version
of the utility `virt-customize` (libguestfs 1.34).
A whole set of Workload and Orchestrator images were created and fully
tested through regression (real application) test. Will add
documentation on the wiki as soon as this commit is merged.