Use glanceclient to access OpenStack image API
When creating VMs in parallel, cbtool creates novaclient client object
in main thread and invokes methods on that object in multiple worker
threads. This approach could cause random failures because novaclient
is not thread-safe. A main issue is that novaclient provides methods
to access image and network APIs and those methods change client
object's service_type attribute, which is one of the components used
to generate request URL. So for example, if a worker thread calls
image API when there are other threads calling compute APIs, the latter
would fail. I observed the issue consistently when I ran SPECcloud's
yscb baseline script (I replaced its cbtool with the latest vesion)
on systems which had latest version of novaclient installed.

The change works around the issue by using glanceclient to call image
API. With this change the issue is gone on my systems.