This is (hopefully) the last of "large" (but again, painfully tested) Pull Requests
1 - Quasi-Universal VPN support
2 - Workload Image building directly from CLI/API
3 - Workload deployment using unconfigured images
4 - Parallel Libvirt Manager (PLM) re-factored
5 - The CLI/API call "shell" can now execute arbitrary code on instances
6 - Bugfixes for multiple cloud adapters

1 - Quasi-Universal VPN support: if a cloud supports cloud-init, its
instances can now leverage the VPN connectivity technique already
built-in on the tool. Three important aspects of this new feature:

a) there is now enough code on the "real multicloud regression" to
actually test VPN support for all clouds

b) the (VM_DEFAULTS global object) option known as VPN_ONLY has been
removed (it has shown itself to be thoroughly irrelevant, in terms of
scalability) : from now on, simply setting the (also on VM_DEFAULTS
global object) USE_VPN_IP=$True should suffice to activate VPN support

c) when VPN support is requested, the cloud-init file is already
populated with "openvpn" and "redis-tools" packages

2 - Workload Image building directly from CLI/API: initial support for
it was introduced a while back, but full use of this feature has
always been limited due to the fact that not all workloads map nicely
to a single image. An example: we us a single image - and thus a single
Dockerfile - for all YCSB workloads, be it MongoDB, Redis and Cassandra.
With the updates made here, CBTOOL now is able to properly map AI types
to Dockerfiles, automatically extracting the correct paramaters to call
the automated installer (just remember, the automated installer parses
the Dockerfiles to figure which commands to execute).

This means that a CLI/API call such as `vmattach
check:<BLANKIMG>:<DEFUSER>:nullworkload:build`
will result in the execution of the automated installer (i.e.,
`~/cbtool/install`) with
the appropriate parameters (i.e., `--role workload --wks nullworkload`).

To further illustrate the somewhat confused point made above: if a
CLI/API call such as
`vmattach check:<BLANKIMG>:<DEFUSER>:cassandra_ycsb:build` is issued,
then the paramaters
for the invocation of the automated installer will be `-role workload
--wks nullworkload,redis_ycsb,mongo_ycsb,cassandra_ycsb`.

After the instance image creation is complete, the user can proceed to
capture it with
the `vmcapture` call

3 - Workload deployment using unconfigured images: the very same
mechanism explained above allows the user to deploy fully functional
workloads on top of "blank" images. For instance, a CLI/API call such
as `aiattach build:<BLANKIMG>:nullworkload` would attempt to execute the
automated installer with the correct paramaters (i.e.,
, `--role workload --wks nullworkload`) before proceeding to the
regular Application Instance deployment sequence (i.e., generic
scripts, application-specific scripts and load manager startup).

4 - Parallel Libvirt Manager (PLM) re-factored: it is now significantly
simpler to use it against a set of (minimally configured) libvirtd
hosts.

5 - The CLI/API call "shell" can now execute arbitrary code on
instances: an illustrative example, used on the "real multicloud
multicloud regression test": a call such as `shell vm_1 sudo fdisk`
is used to make sure that the additional data volumes created during the
test are actually attached to the instance.

6 - Bugfixes for multiple cloud adapters