Parallel Libvirt Manager (PLM) adapter re-factored
It is now significantly simpler to use it against a set of (minimally
configured) libvirtd hosts (basically, just a few commands on a
"pristine" host, as will be explained on the documentation). This,
coupled together with the (virt-customize-based) method to build
full VM (qcow2) images already in place (scripts on "kvm-qemu" dir)
should allow an user to quickly deploy all workloads on a set of hosts.