"Step"-through support for the load manager
This feature allows you do make the load manager pause after each
load ID is complete.

It allows you to "step" through the execution of the load manager
via the API in such a way that the client explicitly controls when
the load manager moves to the next load iteration.

This has the effect of allowing you to re-use the same virtual machine
for more exploratory tests without making a new VM.

The way the feature works is that you have two options:

1. Pause after initial AI attachment.
2. AND/OR Pause after each load ID has completed.

In either case, by issuing polling for the state field via the API,
you will know when the load manager has paused.

Subsequently, a client than then issue "statealter.... attached" to
instruct the load manager to unpause and perform another iteration.

Additionally, we have to query Redis (within the VM) to do to this,
so this required changes in the way we handle "controllable" vs.
"scalable" mode, but it seems like a small price to pay.

We've observed no problem in scalability so far.