Restoration of remote volume support
Until now, attachment of remote volumes for all of the public cloud providers
was pretty much broken.

This does several things to enhance it:

1. It adds support for controlling the exact volume type of the root volume
   AND the remote volume via configuration file.

2. It makes attachment and cleanup work on the clouds.

In addition to the use of `CLOUD_VV_TYPE`, there is now a `CLOUD_RV_TYPE`
(note the `R` instead of the `V` which indicates "root").