A tool to plot graphs for host and VM OS metric data
The change introduces a new graph plotting tool. It read host and VM
OS metric data from CSV files generated by monextract command, save
them into RRDtool database and use RRDtool graph command to plot graphs.
The tool is unique in that it doesn't hardcode metrics and graphs, but
provides a config file in YAML format for user to define them. A default
config file is shipped with the tool as an example.

To run the tool, user need to enable host and VM metric collecting and
run monextract to save metric data into CSV files. Then run the tool:

  $ python rrdtool.py <experiment_dir>

The command will use the default config file to generate graphs in the
experiment_dir. For more details about how it works and how to config it,
please read the file header comments in the tool script and the example
config file.